Flattened config for easier defaults
