+++ src/migrator/linux/boot_manager_impl/u_boot_manager.rs
@@ -15,7 +15,7 @@ use crate::{
     common::{
         boot_manager::BootManager,
         call,
-        config::migrate_config::UEnvStrategy,
+        config::UEnvStrategy,
         is_balena_file,
         migrate_info::MigrateInfo,
         path_append,
@@ -868,7 +868,7 @@ impl BootManager for UBootManager {
         debug!("can_migrate: entered");
 
         // determine flash device - either from config or root device
-        let flash_device = if let Some(dev_path) = config.debug.get_force_flash_device() {
+        let flash_device = if let Some(dev_path) = config.get_force_flash_device() {
             warn!("Config forces use of flash device '{}'", dev_path.display());
             BlockDevice::from_device_path(dev_path)?
         } else {
